Add separation of concern with MVC achitecture
