+++ routes/tourRoute.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const tourControllers = require('../controller/tourController');
+const { getAllTours, createTour, getOneTour, updateTour, deleteTour } =
+  tourControllers;
+// TOURS ROUTES
+const route = express.Router();
+route.route('/').get(getAllTours).post(createTour);
+route.route('/:id').get(getOneTour).patch(updateTour).delete(deleteTour);
+
+module.exports = route;
